refactor: centralize admin runtime sources
