+++ rustfs/src/admin/console.rs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 use crate::admin::handlers::health::{HealthProbe, build_health_response_parts, collect_dependency_readiness};
+use crate::admin::runtime_sources::resolve_oidc_handle;
 use crate::app::admin_usecase::DefaultAdminUsecase;
-use crate::app::context::resolve_oidc_handle;
 use crate::license::has_valid_license;
 use crate::server::has_path_prefix;
 use crate::server::{
