+++ rustfs/src/admin/handlers/kms_dynamic.rs
@@ -17,7 +17,7 @@
 use super::super::{read_admin_config, save_admin_config};
 use crate::admin::auth::validate_admin_request;
 use crate::admin::router::{AdminOperation, Operation, S3Router};
-use crate::app::context::{
+use crate::admin::runtime_sources::{
     resolve_kms_runtime_service_manager, resolve_object_store_handle, resolve_or_init_kms_runtime_service_manager,
 };
 use crate::auth::{check_key_valid, get_session_token};
