+++ rustfs/src/admin/handlers/trace.rs
@@ -14,7 +14,7 @@
 
 use super::super::PeerRestClient;
 use crate::admin::router::Operation;
-use crate::app::context::resolve_endpoints_handle;
+use crate::admin::runtime_sources::resolve_endpoints_handle;
 use http::StatusCode;
 use hyper::Uri;
 use matchit::Params;
