+++ rustfs/src/admin/mod.rs
@@ -21,6 +21,7 @@ mod plugin_contract;
 #[allow(dead_code)]
 pub(crate) mod route_policy;
 pub mod router;
+pub(crate) mod runtime_sources;
 pub mod service;
 pub mod site_replication_identity;
 pub mod utils;
